CI - Handle test panics without SIGTERM (#2079)
* Added a new `fluvio-test` `self_check` for self-diagnostic usage
* Log collections additionally occur when workflow is cancelled instead of just on failure.
* Handle `fluvio-test` panic with file-based flags to compensate for Github runner signal handling changes
* Minor fix to `action-slack` to hopefully fix missing fields in Slack messages

Resolves #2042

+++ crates/fluvio-test/src/main.rs
@@ -1,4 +1,7 @@
-use std::process::exit;
+use std::process::{self, exit};
+use std::env;
+use std::path::Path;
+use std::fs::{remove_file, File};
 use structopt::StructOpt;
 use fluvio::Fluvio;
 use fluvio_test_util::test_meta::{BaseCli, TestCase, TestCli, TestOption};
@@ -18,6 +21,8 @@ use nix::sys::signal::{kill, Signal};
 // This is important for `inventory` crate
 #[allow(unused_imports)]
 use fluvio_test::tests as _;
+
+const CI_FAIL_FLAG: &str = "/tmp/CI_FLUVIO_TEST_FAIL";
 
 fn main() {
     let option = BaseCli::from_args();
@@ -64,14 +69,39 @@ fn main() {
             duration: panic_timer.duration(),
             ..Default::default()
         };
+
         //run_block_on(async { cluster_cleanup(panic_options.clone()).await });
-        eprintln!("Test panicked: {:#?}", panic_info);
-        eprintln!("{}", test_result);
+        println!("{:#?}", panic_info);
+
+        if env::var("CI").is_err() {
+            println!("{}", test_result);
+        }
     }));
 
+    let parent_process_id: u32 = std::process::id();
+
     let test_result = run_test(option.environment.clone(), test_opt, test_meta);
-    cluster_cleanup(option.environment);
-    println!("{}", test_result);
+
+    // If parent process, we want to
+    // * set the exit code
+    // * print test results
+    // * cleanup cluster
+    if process::id() == parent_process_id {
+        cluster_cleanup(option.environment);
+        print_results(parent_process_id, test_result.clone());
+
+        if test_result.success {
+            exit(0)
+        } else {
+            exit(1)
+        }
+    }
+}
+
+fn print_results(parent_process_id: u32, results: TestResult) {
+    if process::id() == parent_process_id {
+        println!("{}", results)
+    }
 }
 
 fn run_test(
@@ -90,15 +120,35 @@ fn run_test(
     }));
 
     // If we've panicked from the test, we need to terminate all the child processes too to stop the test
-    match test_result {
-        Ok(r) => r.unwrap(),
+    let test_result = match test_result {
+        Ok(r) => {
+            let mut res = r.unwrap();
+            if Path::new(CI_FAIL_FLAG).exists() {
+                remove_file(CI_FAIL_FLAG).unwrap();
+                res.success = false;
+            }
+            res
+        }
         Err(_) => {
             // nix uses pid 0 to refer to the group process, so reap the child processes
             let pid = Pid::from_raw(0);
-            kill(pid, Signal::SIGTERM).expect("Unable to kill test process");
-            exit(1);
+
+            // CI uses SIGTERM to report if jobs are cancelled
+            // so we need to report failure a little differently
+            if env::var("CI").is_ok() {
+                // Create a file for CI to look for, bc using signals causes issues
+                // Since we don't need to clean our environment, terminating child proc less important
+                if !Path::new(CI_FAIL_FLAG).exists() {
+                    let _ = File::create(CI_FAIL_FLAG).unwrap();
+                }
+            } else {
+                kill(pid, Signal::SIGTERM).expect("Unable to kill test process");
+            }
+            TestResult::default()
         }
-    }
+    };
+
+    test_result
 }
 
 fn cluster_cleanup(option: EnvironmentSetup) {
